requirements.txt
flask==1.1.2
keras==2.8.0
matplotlib==3.3.4
numpy==1.20.2
pillow==8.1.2
pyyaml==5.4.1
seaborn==0.11.1
tensorflow==2.8.0
tensorflow-hub==0.11.0
werkzeug==1.0.1
tf_slim==1.1.0
